legger til logikk for visning av alertboks for omstillingstoenad og gjenlevende
