--- src/components/OmUfoeregrad/OmUfoeretrygd.module.scss
@@ -1,42 +1,0 @@
-@import '../../scss/variables';
-
-.wrapper {
-  composes: frame from '../../scss/modules/frame.module.scss';
-
-  &Card {
-    padding: var(--a-spacing-4) var(--a-spacing-6);
-
-    @media (min-width: $a-breakpoint-lg) {
-      padding: var(--a-spacing-6) 0;
-      max-width: $container-max-width;
-    }
-  }
-}
-
-.ingress {
-  display: inline;
-
-  @media (min-width: $a-breakpoint-lg) {
-    font-size: var(--a-font-size-xlarge);
-    line-height: var(--a-font-line-height-xlarge);
-  }
-}
-
-.highlighted {
-  display: inline;
-  font-weight: var(--a-font-weight-bold);
-
-  @media (min-width: $a-breakpoint-lg) {
-    font-size: var(--a-font-size-xlarge);
-    line-height: var(--a-font-line-height-xlarge);
-  }
-}
-
-.readmore {
-  margin-top: var(--a-spacing-1);
-}
-
-.alert {
-  margin-top: var(--a-spacing-2);
-  margin-bottom: var(--a-spacing-2);
-}
