--- src/components/OmUfoeregrad/index.ts
@@ -1,1 +1,0 @@
-export { OmUfoeretrygd } from './OmUfoeretrygd'
