+++ src/translations/en.ts
@@ -144,6 +144,7 @@ const translations = {
   'tidligstmuliguttak.readmore_title': 'MANGLER_TEKST',
   'tidligstmuliguttak.readmore_ingress.optional': 'MANGLER_TEKST',
   'tidligstmuliguttak.readmore_ingress.enkelt': 'MANGLER_TEKST',
+  'tidligstmuliguttak.info_omstillinngsstoenad_og_gjenlevende': 'MANGLER_TEKST',
   'tidligstmuliguttak.info_afp': 'MANGLER_TEKST',
   'tidligstmuliguttak.error': 'MANGLER_TEKST',
   'omufoeretrygd.hel.ingress': 'MANGLER_TEKST',
